requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==1.11
django-betterforms==1.2
django-crispy-forms==1.6.1
django-filter==1.0.2
django-tables2==2.0.3
django-widget-tweaks==1.4.1
djangoajax==2.3.5
djangorestframework==3.9.3
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2019.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
